# Conflicts:
#	SPring_Cloud_Group/spring_cloud_uc_server/pom.xml
#	SPring_Cloud_Group/spring_cloud_uc_server/src/main/java/com/rjpa/controller/IndexMvcController.java
#	SPring_Cloud_Group/spring_cloud_uc_server/src/main/java/com/rjpa/controller/UserRoleMvcController.java
#	SPring_Cloud_Group/spring_cloud_uc_server/src/main/resources/templates/login.html

+++ SPring_Cloud_Group/spring_cloud_uc_server/src/main/java/com/rjpa/service/IndexMvcService.java
@@ -11,7 +11,7 @@ public interface IndexMvcService {
      *
      * @return
      */
-    public Result getAdminMenusRights(long parentid);
+    public Result getAdminMenusRights(int parentid);
     public Result getAllAdminMenusRights();
     /**
      * 获取全部系统操作权限名称链接，adminPermission
